refactor(tasks): host the pool handoff and unpark waker as shared helpers
Consolidate the pool-to-poll family onto ecosystem crates. nectar-tasks now
hosts the oneshot pool handoff (a panic-catching `catch_unwind` submit seam
onto the rayon pool, behind `rayon`) and the thread-unpark waker (behind
`std`), the single canonical homes for both.

Migrate the file split handoff and the postage-issuer blocking bridge onto
them: the split handoff module re-exports the shared `submit`/`Handoff`, and
the bridge re-exports the shared `unpark_current`. Drop file's now-unused
`futures-channel` and `rayon` direct dependencies; the `rayon` feature routes
through `nectar-tasks/rayon`.

Semantics are unchanged: a panicked job still reads as a dropped sender, and
the interleaving tests move to the handoff's new home.
