+++ crates/file/Cargo.toml
@@ -20,14 +20,12 @@ workspace = true
 [dependencies]
 arbitrary = { workspace = true, optional = true }
 bytes = { workspace = true, optional = true }
-futures-channel = { workspace = true, optional = true }
 futures-util = { workspace = true, optional = true }
 nectar-clock = { workspace = true, optional = true }
 nectar-kernel = { workspace = true }
 nectar-marker = { workspace = true }
 nectar-primitives = { workspace = true, optional = true }
 nectar-tasks = { workspace = true, optional = true }
-rayon = { workspace = true, optional = true }
 thiserror = { workspace = true }
 tokio = { workspace = true, features = ["sync"], optional = true }
 zeroize = { workspace = true, optional = true }
@@ -76,7 +74,7 @@ tokio = [ "dep:nectar-tasks", "dep:tokio", "nectar-tasks/tokio", "std" ]
 # Pool leaf hashing: the split hash window and the read-at ingest. Implies
 # `std`; rayon on wasm32 or under `unsync` is a compile error, never a
 # silent fallback.
-rayon = [ "dep:futures-channel", "dep:rayon", "std" ]
+rayon = [ "dep:nectar-tasks", "nectar-tasks/rayon", "std" ]
 
 # The shared malformed-walk oracle, the structural-regime body generators
 # and the stable seed replays. Never reachable from `default` or any
