+++ crates/tasks/README.md
@@ -1,6 +1,6 @@
 # nectar-tasks
 
-Object-safe task spawn seam: a `Spawn` trait over boxed unit futures with an abort-on-drop `TaskHandle`, a tokio spawner behind the `tokio` feature and a browser spawner behind the `wasm` feature. Thread-safety bounds relax on single-threaded targets via `nectar-marker`.
+Object-safe task spawn seam: a `Spawn` trait over boxed unit futures with an abort-on-drop `TaskHandle`, a tokio spawner behind the `tokio` feature and a browser spawner behind the `wasm` feature. Thread-safety bounds relax on single-threaded targets via `nectar-marker`. Blocking bridges get a thread-unpark waker behind `std` and a panic-catching pool handoff over a oneshot channel behind `rayon`.
 
 Part of the [nectar](https://github.com/nxm-rs/nectar) workspace, a collection of low-level Ethereum Swarm primitives in Rust. See the [workspace README](https://github.com/nxm-rs/nectar) for the full crate list and project context.
